+++ RELEASE_NOTES.md
@@ -1,2 +1,7 @@
+** Release Notes
+
+*** Development Version
+
 - Initializing context and tools for the session.
 - Removed gg.py and gui_files/ directory.
+- Renaming project and its files as: gpaw-tools -> Pint, gpawsolve.py -> dftsolve.py, and asapsolve -> mdsolve.py